+++ CHANGELOG.md
@@ -13,13 +13,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed: for any bug fixes.
 - Security: in case of vulnerabilities
 
-## [0.1.0] - Unreleased
+## [0.1.0](https://github.com/SergeyGadzhilov/SKVM/releases/tag/0.1.0) - 2024-02-04
 ### Added
 - Status bar into the MainWindow [#76](https://github.com/SergeyGadzhilov/SKVM/issues/76)
 - Notifications system [#64](https://github.com/SergeyGadzhilov/SKVM/issues/64)
 - Notifications about new version [#61](https://github.com/SergeyGadzhilov/SKVM/issues/61)
 
-## [0.0.3] - (https://github.com/SergeyGadzhilov/SKVM/releases/tag/0.0.3) - 2024-01-29
+## [0.0.3](https://github.com/SergeyGadzhilov/SKVM/releases/tag/0.0.3) - 2024-01-29
 ### Added
 - Logs to the sidebar [#63](https://github.com/SergeyGadzhilov/SKVM/issues/63)
 - Setting to the sidebar [#62](https://github.com/SergeyGadzhilov/SKVM/issues/62)